refactor the network logic into its own hook
