+++ src/App.tsx
@@ -14,6 +14,7 @@ import ScreenPokedex from "./ScreenPokedex";
 import ScreenPokedexHelp from "./ScreenPokedexHelp";
 import ScreenWeaknessCoverage from "./ScreenWeaknessCoverage";
 import { PUBLIC_PATH } from "./settings";
+import { useFetchJSON } from "./useFetchJSON";
 import { useLanguage } from "./useLanguage";
 import { useTheme } from "./useTheme";
 import { useUpdateSW } from "./useUpdateSW";
@@ -39,12 +40,15 @@ const tabClass = classNames([
 const tabClassActive = classNames(["fg1 bottom-border-thick-current"]);
 
 export default function App() {
+  // Service worker
   const {
     needRefresh: [needRefresh, setNeedRefresh],
     offlineReady: [offlineReady, setOfflineReady],
     updateServiceWorker,
   } = useRegisterSW();
   useUpdateSW();
+
+  // State...
   const [defenseParams, setDefenseParams] = React.useState("");
   const [offenseParams, setOffenseParams] = React.useState("");
   const [pokedexParams, setPokedexParams] = React.useState("");
@@ -54,44 +58,41 @@ export default function App() {
     CoverageType[]
   >([]);
   const [AllPokemon, setAllPokemon] = React.useState<Pokemon[]>([]);
-  const [attemptTime, setAttemptTime] = React.useState(Date.now());
+
   const [language] = useLanguage();
-  React.useEffect(() => {
-    async function load() {
-      const filename = "data-pkmn.json";
-      try {
-        const bigURL = new URL(filename, PUBLIC_PATH).href;
-        const resp = await fetch(bigURL);
-        const allPokemon: Pokemon[] = await resp.json();
-        const fallbackCoverageTypes = allPokemon.map<CoverageType>((pkmn) => {
-          const name = formatPokemonName({
-            speciesName: pkmn.speciesNames[language],
-            formName: pkmn.formNames[language],
-          });
-          const number = String(pkmn.number);
-          const types = pkmn.types;
-          return { number, name, types };
-        });
-        setIsLoading(false);
-        setCoverageTypes(fallbackCoverageTypes);
-        setFallbackCoverageTypes(fallbackCoverageTypes);
-        setAllPokemon(allPokemon);
-      } catch (err) {
-        console.warn(`Failed to download ${filename}`, err);
-        const retryDelay = 60 * 1000;
-        // Retry every minute until the JSON finishes downloading
-        setTimeout(() => {
-          setAttemptTime(Date.now());
-        }, retryDelay);
-      }
-    }
-    load();
-  }, [attemptTime]);
+
+  // Theme stuff
   const [theme] = useTheme();
   const isDarkMode = useMediaQuery("(prefers-color-scheme: dark)");
   const isDark = theme === "dark" || (theme === "auto" && isDarkMode);
   const themeColor = isDark ? "hsl(350, 70%, 40%)" : "hsl(357, 97%, 46%)";
   const themeAuto = isDarkMode ? "dark" : "light";
+
+  // Load Pokédex JSON
+  const jsonURL = new URL("data-pkmn.json", PUBLIC_PATH).href;
+  const allPokemonResponse = useFetchJSON<Pokemon[]>(jsonURL);
+
+  // Fallback coverage types
+  React.useEffect(() => {
+    if (allPokemonResponse.type !== "done") {
+      return;
+    }
+    const allPokemon = allPokemonResponse.data;
+    const fallbackCoverageTypes = allPokemon.map<CoverageType>((pkmn) => {
+      const name = formatPokemonName({
+        speciesName: pkmn.speciesNames[language],
+        formName: pkmn.formNames[language],
+      });
+      const number = String(pkmn.number);
+      const types = pkmn.types;
+      return { number, name, types };
+    });
+    setIsLoading(false);
+    setCoverageTypes(fallbackCoverageTypes);
+    setFallbackCoverageTypes(fallbackCoverageTypes);
+    setAllPokemon(allPokemon);
+  }, [allPokemonResponse, language]);
+
   return (
     <HelmetProvider>
       <div className="flex-auto">
